# swagger.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "info": {
    "title": "MRCS Question Extractor API",
    "description": "AI-powered medical exam question extraction from PDF documents",
    "version": "1.0.0",
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
...
    "/intakes": {
      "get": {
        "summary": "Get all intakes",
        "description": "Get all intakes with pagination and filtering",
        "tags": [
          "intakes"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDto__type.o1"
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "intakes",
          "findAll"
        ],
        "x-samchon-controller": "IntakesController"
      },
      "post": {
        "summary": "Create a new intake",
        "description": "Create a new intake",
        "tags": [
          "intakes"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIntakeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "type": {
                      "oneOf": [
                        {
                          "const": "JANUARY"
                        },
                        {
                          "const": "APRIL_MAY"
                        },
                        {
                          "const": "SEPTEMBER"
                        }
                      ]
                    },
                    "questionCount": {
                      "type": "number"
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "name",
                    "id",
                    "displayName",
                    "type",
                    "questionCount",
                    "isActive",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "intakes",
          "create"
        ],
        "x-samchon-controller": "IntakesController"
      }
    },
...
    "/intakes/{id}": {
      "get": {
        "summary": "Get intake by ID",
        "description": "Get intake by ID",
        "tags": [
          "intakes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "id": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "type": {
                          "oneOf": [
                            {
                              "const": "JANUARY"
                            },
                            {
                              "const": "APRIL_MAY"
                            },
                            {
                              "const": "SEPTEMBER"
                            }
                          ]
                        },
                        "questionCount": {
                          "type": "number"
                        },
                        "isActive": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "name",
                        "id",
                        "displayName",
                        "type",
                        "questionCount",
                        "isActive",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "intakes",
          "findById"
        ],
        "x-samchon-controller": "IntakesController"
      },
      "put": {
        "summary": "Update an intake",
        "description": "Update an intake",
        "tags": [
          "intakes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIntakeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "id": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "type": {
                          "oneOf": [
                            {
                              "const": "JANUARY"
                            },
                            {
                              "const": "APRIL_MAY"
                            },
                            {
                              "const": "SEPTEMBER"
                            }
                          ]
                        },
                        "questionCount": {
                          "type": "number"
                        },
                        "isActive": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "name",
                        "id",
                        "displayName",
                        "type",
                        "questionCount",
                        "isActive",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "intakes",
          "update"
        ],
        "x-samchon-controller": "IntakesController"
      },
      "delete": {
        "summary": "Delete an intake",
        "description": "Delete an intake (only custom intakes can be deleted)",
        "tags": [
          "intakes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "content": {
              "application/json": {}
            }
          }
        },
        "x-samchon-accessor": [
          "intakes",
          "$delete"
        ],
        "x-samchon-controller": "IntakesController"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PaginatedResponseDto__type.o1": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "id": {
                  "type": "string"
                },
                "displayName": {
                  "type": "string"
                },
                "type": {
                  "oneOf": [
                    {
                      "const": "JANUARY"
                    },
                    {
                      "const": "APRIL_MAY"
                    },
                    {
                      "const": "SEPTEMBER"
                    }
                  ]
                },
                "questionCount": {
                  "type": "number"
                },
                "isActive": {
                  "type": "boolean"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "name",
                "id",
                "displayName",
                "type",
                "questionCount",
                "isActive",
                "createdAt",
                "updatedAt"
              ]
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "limit": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "totalPages": {
                "type": "number"
              },
              "hasNextPage": {
                "type": "boolean"
              },
              "hasPrevPage": {
                "type": "boolean"
              }
            },
            "required": [
              "page",
              "limit",
              "total",
              "totalPages",
              "hasNextPage",
              "hasPrevPage"
            ]
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "pagination",
          "success"
        ]
      },
      "CreateIntakeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "oneOf": [
              {
                "const": "JANUARY"
              },
              {
                "const": "APRIL_MAY"
              },
              {
                "const": "SEPTEMBER"
              }
            ]
          },
          "displayName": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "type",
          "displayName"
        ]
      },
      "UpdateIntakeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "oneOf": [
              {
                "const": "JANUARY"
              },
              {
                "const": "APRIL_MAY"
              },
              {
                "const": "SEPTEMBER"
              }
            ]
          },
          "displayName": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": []
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}